requirements.txt
tabulate
dummypackage
ulauncher-api>=2.0.0

